youtube-scraper-app/requirements.txt
# YouTube Scraper Application - Dependencies

# GUI Framework
PyQt6==6.6.1
PyQt6-Qt6==6.6.1

# YouTube Integration
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
yt-dlp==2023.12.30

# Database
psycopg2-binary==2.9.9

# Async & Parallel Processing
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2

# Data Processing
pandas==2.1.4

# Text Processing & ML
scikit-learn==1.3.2
nltk==3.8.1

# Logging
colorlog==6.8.0

# Progress Bars
tqdm==4.66.1

